Refactored utils and imports
